requirements.txt
# Core
discord.py[voice]>=2.3.0
python-dotenv>=1.0.0

# Audio
yt-dlp>=2024.1.0
PyNaCl>=1.5.0

# HTTP (lyrics)
aiohttp>=3.9.0

# Spotify (optional — only needed if SPOTIFY_CLIENT_ID is set in .env)
spotipy>=2.23.0
